library template + script
