--- librarytemplate/src/PlcTemplateConnector/PlcTemplateConnector.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net48</TargetFramework>    
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <PackageReference Include="Inxton.Package.Vortex.Essentials" Version="1.10.0-nighlty.436" />
-  </ItemGroup>  
-</Project>
