--- librarytemplate/src/Wpf/PlcTemplate.Wpf/PlcTemplate.Wpf.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net48</TargetFramework>
-    <UseWPF>true</UseWPF>  
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Inxton.Package.Vortex.Essentials.Wpf" Version="1.10.0-nighlty.436" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\PlcTemplateConnector\PlcTemplateConnector.csproj" />
-  </ItemGroup>
-
-</Project>
