--- librarytemplate/tests/PlcUnitTests/PlcTemplateUnitTests.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="nunit" Version="3.12.0" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />   
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\PlcTemplateConnector\PlcTemplateConnector.csproj" />
-  </ItemGroup>
-
-</Project>
