--- librarytemplate/tests/Sandbox.PlcTemplate.Wpf/App.xaml
@@ -1,124 +1,0 @@
-﻿<Application
-    x:Class="Sandbox.PlcTemplate.Wpf.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:Sandbox.PlcTemplate.Wpf"
-    StartupUri="MainWindow.xaml">
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="pack://application:,,,/Vortex.Presentation.Styling.Wpf;component/VortexStyle.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-
-            <Color x:Key="MtsGreenColor">#FF20A024</Color>
-            <Color x:Key="MtsYellowColor">#FFFFD917</Color>
-            <Color x:Key="MtsRedColor">#FFE54E31</Color>
-            <Color x:Key="MtsLightBlueColor">#FF2195F2</Color>
-            <Color x:Key="MtsBlueColor">#FF2043A6</Color>
-            <Color x:Key="MtsLightGrayColor">#FFE6E6E8</Color>
-            <Color x:Key="MtsGrayColor">#FF515560</Color>
-            <Color x:Key="MtsDarkGrayColor">#FF34373F</Color>
-            <Color x:Key="MtsBlackColor">#FF2D2D30</Color>
-            <Color x:Key="MtsOrangeColor">#FFE54E31</Color>
-
-            <SolidColorBrush x:Key="MtsGreen" Color="{DynamicResource MtsGreenColor}" />
-            <SolidColorBrush x:Key="MtsYellow" Color="{DynamicResource MtsYellowColor}" />
-            <SolidColorBrush x:Key="MtsRed" Color="{DynamicResource MtsRedColor}" />
-            <SolidColorBrush x:Key="MtsLightBlue" Color="{DynamicResource MtsLightBlueColor}" />
-            <SolidColorBrush x:Key="MtsBlue" Color="{DynamicResource MtsBlueColor}" />
-            <SolidColorBrush x:Key="MtsLightGray" Color="{DynamicResource MtsLightGrayColor}" />
-            <SolidColorBrush x:Key="MtsGray" Color="{DynamicResource MtsGrayColor}" />
-            <SolidColorBrush x:Key="MtsDarkGray" Color="{DynamicResource MtsDarkGrayColor}" />
-            <SolidColorBrush x:Key="MtsBlack" Color="{DynamicResource MtsBlackColor}" />
-            <SolidColorBrush x:Key="MtsOrange" Color="{DynamicResource MtsOrangeColor}" />
-
-            <Color x:Key="OnMtsGreenColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsYellowColor">#FF2D2D30</Color>
-            <Color x:Key="OnMtsRedColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsLightBlueColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsBlueColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsLightGrayColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsGrayColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsDarkGrayColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsBlackColor">#FFFFFFFF</Color>
-            <Color x:Key="OnMtsOrangeColor">#FFFFFFFF</Color>
-
-            <SolidColorBrush x:Key="OnMtsGreen" Color="{DynamicResource OnMtsGreenColor}" />
-            <SolidColorBrush x:Key="OnMtsYellow" Color="{DynamicResource OnMtsYellowColor}" />
-            <SolidColorBrush x:Key="OnMtsRed" Color="{DynamicResource OnMtsRedColor}" />
-            <SolidColorBrush x:Key="OnMtsLightBlue" Color="{DynamicResource OnMtsLightBlueColor}" />
-            <SolidColorBrush x:Key="OnMtsBlue" Color="{DynamicResource OnMtsBlueColor}" />
-            <SolidColorBrush x:Key="OnMtsLightGray" Color="{DynamicResource OnMtsLightGrayColor}" />
-            <SolidColorBrush x:Key="OnMtsGray" Color="{DynamicResource OnMtsGrayColor}" />
-            <SolidColorBrush x:Key="OnMtsDarkGray" Color="{DynamicResource OnMtsDarkGrayColor}" />
-            <SolidColorBrush x:Key="OnMtsBlack" Color="{DynamicResource OnMtsBlackColor}" />
-            <SolidColorBrush x:Key="OnMtsOrange" Color="{DynamicResource OnMtsOrangeColor}" />
-
-
-
-            <Color x:Key="InxtonLime">#c4d93f</Color>
-            <Color x:Key="InxtonGray">#5f6371</Color>
-            <Color x:Key="InxtonBlue">#c4d93f</Color>
-            <Color x:Key="InxtonOrange">#ff4b27</Color>
-
-
-            <Color x:Key="PrimaryColor">#c4d93f</Color>
-            <Color x:Key="OnPrimaryColor">White</Color>
-
-            <SolidColorBrush x:Key="Primary" Color="{StaticResource MtsOrangeColor}" />
-            <SolidColorBrush x:Key="OnPrimary" Color="{StaticResource OnPrimaryColor}" />
-
-            <Color x:Key="SecondaryColor">#5f6371</Color>
-            <Color x:Key="OnSecondaryColor">#ffffff</Color>
-            <SolidColorBrush x:Key="Secondary" Color="{StaticResource MtsBlueColor}" />
-            <SolidColorBrush x:Key="OnSecondary" Color="{StaticResource OnSecondaryColor}" />
-
-            <Color x:Key="BackgroundColor">White</Color>
-            <Color x:Key="OnBackgroundColor">Black</Color>
-            <SolidColorBrush x:Key="Background" Color="{StaticResource BackgroundColor}" />
-            <SolidColorBrush x:Key="OnBackground" Color="{StaticResource OnBackgroundColor}" />
-
-            <Color x:Key="SurfaceColor">#f1f1f1</Color>
-            <Color x:Key="OnSurfaceColor">Black</Color>
-            <SolidColorBrush x:Key="Surface" Color="{StaticResource SurfaceColor}" />
-            <SolidColorBrush x:Key="OnSurface" Color="{StaticResource OnSurfaceColor}" />
-
-            <Color x:Key="ErrorColor">#ff4b27</Color>
-            <Color x:Key="OnErrorColor">#ffffff</Color>
-
-            <SolidColorBrush x:Key="Error" Color="{StaticResource MtsOrange}" />
-            <SolidColorBrush x:Key="OnError" Color="{StaticResource OnPrimaryColor}" />
-
-            <Color x:Key="ShadowColor">Black</Color>
-
-            <Color x:Key="AccentColor">#ff4b27</Color>
-            <Color x:Key="OnAccentColor">#ffffff</Color>
-            <SolidColorBrush x:Key="Accent" Color="{StaticResource AccentColor}" />
-            <SolidColorBrush x:Key="OnAccent" Color="{StaticResource OnAccentColor}" />
-
-            <Color x:Key="AlertColor">#ca2f14</Color>
-            <Color x:Key="OnAlertColor">#FFFF00</Color>
-            <SolidColorBrush x:Key="Alert" Color="{StaticResource AlertColor}" />
-            <SolidColorBrush x:Key="OnAlert" Color="{StaticResource OnAlertColor}" />
-
-            <Color x:Key="WarningColor">#ff4b27</Color>
-            <Color x:Key="OnWarningColor">White</Color>
-            <SolidColorBrush x:Key="Warning" Color="{StaticResource MtsYellowColor}" />
-            <SolidColorBrush x:Key="OnWarning" Color="{StaticResource OnWarningColor}" />
-
-            <Color x:Key="SignalOffColor">#5f6371</Color>
-            <Color x:Key="OnSignalOffColor">White</Color>
-            <SolidColorBrush x:Key="SignalOff" Color="{StaticResource MtsGrayColor}" />
-            <SolidColorBrush x:Key="OnSignalOff" Color="{StaticResource OnSignalOffColor}" />
-
-            <Color x:Key="SignalOnColor">#c4d93f</Color>
-            <Color x:Key="OnSignalOnColor">White</Color>
-            <SolidColorBrush x:Key="SignalOn" Color="{StaticResource SignalOnColor}" />
-            <SolidColorBrush x:Key="OnSignalOn" Color="{StaticResource OnSignalOnColor}" />
-
-            <Color x:Key="IntermediateColor">SteelBlue</Color>
-            <Color x:Key="OnIntermediateColor">White</Color>
-
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
