--- librarytemplate/tests/Sandbox.PlcTemplate.Wpf/App.xaml.cs
@@ -1,22 +1,0 @@
-﻿using PlcTemplate;
-using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace Sandbox.PlcTemplate.Wpf
-{
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application
-    {
-        public App()
-        {
-            Entry.PlcTemplatePlc.Connector.BuildAndStart().ReadWriteCycleDelay = 75;
-        }
-    }
-}
