+++ src/librarytemplate/PlcTemplate.slnf
@@ -1,0 +1,13 @@
+{
+  "solution": {
+    "path": "..\\..\\TcOpen.sln",
+    "projects": [
+      "src\\librarytemplate\\src\\PlcTemplateConnector\\PlcTemplateConnector.csproj",
+      "src\\librarytemplate\\src\\Wpf\\PlcTemplate.Wpf\\PlcTemplate.Wpf.csproj",
+      "src\\librarytemplate\\src\\XAE\\XAE\\XAEPlcTemplate.tsproj",
+      "src\\librarytemplate\\tests\\PlcTemplateTestsConnector\\PlcTemplateTestsConnector.csproj",
+      "src\\librarytemplate\\tests\\PlcUnitTests\\PlcTemplateUnitTests.csproj",
+      "src\\librarytemplate\\tests\\Sandbox.PlcTemplate.Wpf\\Sandbox.PlcTemplate.Wpf.csproj"
+    ]
+  }
+}
