+++ src/librarytemplate/src/PlcTemplateConnector/Properties/AssemblyInfo.cs
