+++ src/librarytemplate/src/PlcTemplateConnector/Properties/Localizations.resx
