+++ src/librarytemplate/src/Wpf/PlcTemplate.Wpf/AssemblyInfo.cs
