+++ src/librarytemplate/src/Wpf/PlcTemplate.Wpf/PlcTemplate.Wpf.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+    <TargetFramework>net48</TargetFramework>
+    <UseWPF>true</UseWPF>  
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\TcoCore\src\TcoCore.Wpf\TcoCore.Wpf.csproj" />
+    <ProjectReference Include="..\..\PlcTemplateConnector\PlcTemplateConnector.csproj" />
+  </ItemGroup>
+
+</Project>
