+++ src/librarytemplate/src/Wpf/PlcTemplate.Wpf/_Sample/SampleComponentManualView.xaml
@@ -1,0 +1,43 @@
+﻿<UserControl
+    x:Class="PlcTemplate.SampleComponentManualView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:PlcTemplate"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:plctemplate="clr-namespace:PlcTemplate;assembly=PlcTemplateConnector"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d">
+    <UserControl.DataContext>
+        <plctemplate:SampleComponent />
+    </UserControl.DataContext>
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <GroupBox
+            Grid.Row="0"
+            DataContext="{Binding _addNumberOne}"
+            Header="{Binding AttributeName}">
+            <Slider
+                Maximum="{Binding InstanceMaxValue}"
+                Minimum="{Binding InstanceMinValue}"
+                Value="{Binding Cyclic}" />
+        </GroupBox>
+        <GroupBox
+            Grid.Row="1"
+            DataContext="{Binding _addNumberTwo}"
+            Header="{Binding AttributeName}">
+            <Slider
+                Maximum="{Binding InstanceMaxValue}"
+                Minimum="{Binding InstanceMinValue}"
+                Value="{Binding Cyclic}" />
+        </GroupBox>
+        <GroupBox Grid.Row="2" Header="{Binding AttributeName}" DataContext="{Binding _addResult}">
+            <TextBlock Text="{Binding Cyclic}" HorizontalAlignment="Center" FontSize="20" />
+        </GroupBox>
+    </Grid>
+</UserControl>
