+++ src/librarytemplate/src/Wpf/PlcTemplate.Wpf/_Sample/fbSampleComponentManualView.xaml.cs
