+++ src/librarytemplate/src/XAE/XAE/PlcTemplate/POUs/MAIN.TcPOU
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TcPlcObject Version="1.1.0.1" ProductVersion="3.1.4024.6">
+  <POU Name="MAIN" Id="{55e5d4d9-4607-4a4e-b397-63e624f66e96}" SpecialFunc="None">
+    <Declaration><![CDATA[PROGRAM MAIN
+VAR
+   _sampleContext : SampleContext;
+END_VAR
+]]></Declaration>
+    <Implementation>
+      <ST><![CDATA[_sampleContext.Run();]]></ST>
+    </Implementation>
+  </POU>
+</TcPlcObject>
