+++ src/librarytemplate/src/XAE/XAE/PlcTemplate/POUs/SampleComponent/SampleTaskFunctionBuffer.TcDUT
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TcPlcObject Version="1.1.0.1" ProductVersion="3.1.4024.6">
+  <DUT Name="SampleTaskFunctionBuffer" Id="{056d4909-fab7-493d-b7d6-9ed7ad026b81}">
+    <Declaration><![CDATA[TYPE SampleTaskFunctionBuffer :
+STRUCT
+	x : LREAL;
+	y : LREAL;
+END_STRUCT
+END_TYPE
+]]></Declaration>
+  </DUT>
+</TcPlcObject>
