+++ src/librarytemplate/src/XAE/XAE/PlcTemplate/POUs/SampleComponent/SampleTaskIteration.TcPOU
@@ -1,0 +1,42 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TcPlcObject Version="1.1.0.1" ProductVersion="3.1.4024.6">
+  <POU Name="SampleTaskIteration" Id="{ff8ab0ff-7eb0-0983-34f1-16de82b1c941}" SpecialFunc="None">
+    <Declaration><![CDATA[FUNCTION_BLOCK SampleTaskIteration EXTENDS TcoCore.TcoTask
+VAR	
+	_x : LREAL;
+	_y : LREAL;
+	_d : LREAL;
+	{attribute addProperty Minimum 100}
+	{attribute addProperty Maximum 1000}
+	_iterations : INT;	
+	_currentIteration : INT;	
+END_VAR]]></Declaration>
+    <Implementation>
+      <ST><![CDATA[IF(Execute()) THEN
+	_y := SIN(_x);
+	_x := _x + _d;		
+	_currentIteration := _currentIteration + 1;	
+	DoneWhen(_currentIteration >= _iterations);
+END_IF]]></ST>
+    </Implementation>
+    <Method Name="Restore" Id="{1d66f1bc-4fc9-0e1d-2089-a2db366a8b62}">
+      <Declaration><![CDATA[(*~
+	<docu>
+		<summary>
+			Restores this instance to the Ready state. Can be called explicitly or from within one of the restore mechanisms.
+			<para>
+				See also <see cref="TcoTask.PlcTcoTask.AutoRestorable()"/>
+			</para>
+		</summary>			
+	</docu>	
+~*)
+METHOD Restore : ITcoRestorable]]></Declaration>
+      <Implementation>
+        <ST><![CDATA[SUPER^.Restore();
+_x := 0.0;
+_y := 0.0;
+_currentIteration := 0;]]></ST>
+      </Implementation>
+    </Method>
+  </POU>
+</TcPlcObject>
