+++ src/librarytemplate/src/XAE/XAE/PlcTemplate/POUs/SampleContext.TcPOU
@@ -1,0 +1,28 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TcPlcObject Version="1.1.0.1" ProductVersion="3.1.4024.6">
+  <POU Name="SampleContext" Id="{17b40f53-8ddf-02a7-0c48-23be770b6c32}" SpecialFunc="None">
+    <Declaration><![CDATA[FUNCTION_BLOCK SampleContext EXTENDS TcoCore.TcoContext
+VAR
+	_sampleComponent : SampleComponent(THIS^);
+END_VAR]]></Declaration>
+    <Implementation>
+      <ST><![CDATA[]]></ST>
+    </Implementation>
+    <Method Name="Main" Id="{b16c1f30-7567-0482-3818-51b562bd8312}">
+      <Declaration><![CDATA[(*~
+	<docu>
+		<summary>
+			Main method of the TcoContext. This is the entry point of any control logic that belongs to this context. 
+			The call of this method is ensured by calling the <c>InstanceName.Run()</c> method, and it must not be called explicitly.
+			This method is abstract, and it must be overridden in derived block.
+		</summary>
+	</docu>
+~*)
+METHOD PROTECTED  Main]]></Declaration>
+      <Implementation>
+        <ST><![CDATA[_sampleComponent.Service();
+_sampleComponent();]]></ST>
+      </Implementation>
+    </Method>
+  </POU>
+</TcPlcObject>
