+++ src/librarytemplate/src/XAE/XAE/PlcTemplate/POUs/fbSampleComponent.TcPOU
