+++ src/librarytemplate/tests/Sandbox.PlcTemplate.Wpf/AssemblyInfo.cs
