+++ src/librarytemplate/tests/Sandbox.PlcTemplate.Wpf/MainWindow.xaml
@@ -1,0 +1,25 @@
+﻿<Window
+    x:Class="Sandbox.PlcTemplate.Wpf.MainWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:Sandbox.PlcTemplate.Wpf"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"  
+    xmlns:vortex="http://vortex.mts/xaml"
+    Width="800"
+    Height="450"
+    mc:Ignorable="d"
+    Title="MainWindow">
+    <Window.DataContext>
+        <local:MainWindowViewModel />
+    </Window.DataContext>
+    <Grid Grid.Row="0" DataContext="{Binding PlcTemplatePlc.MAIN._sampleContext, Mode=OneWay}">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <vortex:RenderableContentControl Grid.Row="0" DataContext="{Binding _sampleComponent._sampleTask, Mode=OneWay}" PresentationType="Base" />
+        <vortex:RenderableContentControl Grid.Row="1" DataContext="{Binding _sampleComponent._sampleTask, Mode=OneWay}" PresentationType="ControlSlim" />
+    </Grid>
+
+</Window>
