+++ src/librarytemplate/tests/Sandbox.PlcTemplate.Wpf/MainWindow.xaml.cs
