+++ src/librarytemplate/tests/Sandbox.PlcTemplate.Wpf/Sandbox.PlcTemplate.Wpf.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>net48</TargetFramework>
+    <UseWPF>true</UseWPF>
+  </PropertyGroup> 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Wpf\PlcTemplate.Wpf\PlcTemplate.Wpf.csproj" />
+  </ItemGroup>
+</Project>
